requirements.txt
pandas>=2.1
pyarrow>=14.0
openpyxl>=3.1
streamlit>=1.31
plotly>=5.18
pyyaml>=6.0
rapidfuzz>=3.6
numpy>=1.26
